♻️ expose use_cases public API and consolidate imports
