+++ jukebox/adapters/inbound/admin/api/current_tag_router.py
@@ -11,11 +11,7 @@ from jukebox.adapters.inbound.admin.api.models import (
     DiscPatchInput,
 )
 from jukebox.domain.entities import CurrentTagStatus, Disc, DiscMetadata, DiscOption
-from jukebox.domain.use_cases.library.add_disc import AddDisc
-from jukebox.domain.use_cases.library.edit_disc import EditDisc
-from jukebox.domain.use_cases.library.get_current_tag_status import GetCurrentTagStatus
-from jukebox.domain.use_cases.library.get_disc import GetDisc
-from jukebox.domain.use_cases.library.remove_disc import RemoveDisc
+from jukebox.domain.use_cases import AddDisc, EditDisc, GetCurrentTagStatus, GetDisc, RemoveDisc
 
 
 def build_current_tag_router(
