+++ jukebox/adapters/inbound/admin/api/discs_router.py
@@ -3,11 +3,7 @@ from pydantic import ValidationError
 
 from jukebox.adapters.inbound.admin.api.models import DiscInput, DiscOutput, DiscPatchInput
 from jukebox.domain.entities import Disc, DiscMetadata, DiscOption
-from jukebox.domain.use_cases.library.add_disc import AddDisc
-from jukebox.domain.use_cases.library.edit_disc import EditDisc
-from jukebox.domain.use_cases.library.get_disc import GetDisc
-from jukebox.domain.use_cases.library.list_discs import ListDiscs
-from jukebox.domain.use_cases.library.remove_disc import RemoveDisc
+from jukebox.domain.use_cases import AddDisc, EditDisc, GetDisc, ListDiscs, RemoveDisc
 
 
 def build_discs_router(
