+++ jukebox/domain/use_cases/__init__.py
@@ -1,4 +1,23 @@
 from .handle_tag_event import HandleTagEvent
+from .library.add_disc import AddDisc
+from .library.edit_disc import EditDisc
+from .library.get_current_tag_status import GetCurrentTagStatus
+from .library.get_disc import GetDisc
+from .library.list_discs import ListDiscs
+from .library.remove_disc import RemoveDisc
+from .library.resolve_tag_id import ResolveTagId
+from .library.search_discs import SearchDiscs
 from .sync_current_tag import SyncCurrentTag
 
-__all__ = ["HandleTagEvent", "SyncCurrentTag"]
+__all__ = [
+    "HandleTagEvent",
+    "AddDisc",
+    "EditDisc",
+    "GetCurrentTagStatus",
+    "GetDisc",
+    "ListDiscs",
+    "RemoveDisc",
+    "ResolveTagId",
+    "SearchDiscs",
+    "SyncCurrentTag",
+]
